[REFACTO] - remove IslandWidget and call HexagonItem directly.
